code/nlq2sparql/pyproject.toml
[tool.poetry]
name = "nlq2sparql"
version = "0.1.0"
description = "Natural Language Query to SPARQL generator for LinkedMusic data lake"
authors = ["LinkedMusic Team"]
readme = "README.md"
packages = [{include = "nlq2sparql"}]

[tool.poetry.dependencies]
python = "^3.9"
google-generativeai = "^0.8.0"
openai = "^1.0.0"
anthropic = "^0.34.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
click = "^8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
nlq2sparql = "nlq2sparql.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
